# src/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: node-swagger-template
# during dev, should point to your local machine.
# Note this field is optional.
host: localhost:10010
# basePath prefixes all resource paths
basePath: /api/v1
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  basicAuth:
    type: basic
security:
  - basicAuth: []
  
paths:
  # /api/v1/movie
  /movie/{id}:
    x-swagger-router-controller: MovieController
    get:
      description: Gets a Movie by id.
      operationId: getMovie
      parameters:
        - name: id
          in: path
          description: The ID of the movie to get.
          required: true
          type: string
          format: ObjectId
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Movie"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Deletes a Movie by id.
      operationId: deleteMovie
      parameters:
        - name: id
          in: path
          description: The ID of the movie to get.
          required: true
          type: string
          format: ObjectId
      responses:
        "204":
          description: Success
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /movie:
    x-swagger-router-controller: MovieController
    get:
      description: Gets all movies created so far.
      operationId: getMovies
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/Movie"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Creates a new movie.
      operationId: postMovie
      parameters:
        - name: body
          in: body
          description: The new Movie object.
          required: true
          schema:
            $ref: "#/definitions/Movie"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/Movie"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  Movie:
    type: object
    required:
      - title
      - plot
      - releaseDate
      - actors
      - duration
      - poster
      - rating
    properties:
      id:
        type: string
        format: ObjectId
      title:
        type: string      
      plot:
        type: string
      releaseDate:
        type: string
        format: date
      actors:
        type: "array"
        items:
          type: string
      duration:
        type: number
      poster:
        type: string
      rating:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string